+++ apps/decodex/src/orchestrator/tests/review_landing_orchestration/landed_lineage_merge/authority_boundaries.rs
@@ -1,0 +1,259 @@
+use crate::orchestrator::{
+	self, StateStore,
+	tests::{
+		self, FakePullRequestReviewStateInspector, FakeTracker,
+		review_landing_classification_review, review_landing_status_support,
+	},
+};
+
+#[test]
+fn reconcile_post_review_orchestration_blocks_admin_merge_for_authority_boundary() {
+	let (temp_dir, config, workflow) = tests::temp_project_layout();
+	let (gh_command_path, invocation_log_path) =
+		tests::install_fake_admin_merge_gh_response(&temp_dir);
+	let config = tests::service_config_with_github_token_env_var_and_command_path(
+		&config,
+		"PATH",
+		&gh_command_path,
+	);
+	let repo_root = config.repo_root().to_path_buf();
+	let issue = review_landing_status_support::post_review_sample_service_owned_issue("In Review");
+	let tracker =
+		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let pr_url = "https://github.com/hack-ink/decodex/pull/173";
+	let merge_subject = r#"{"schema":"decodex/commit/1","summary":"current retained handoff","authority":"PUB-101"}"#;
+	let head_oid =
+		tests::commit_worktree_change(&repo_root, "retained.txt", "ready\n", merge_subject);
+
+	state_store
+		.upsert_worktree("pubfi", &issue.id, "main", &repo_root.display().to_string())
+		.expect("worktree should record");
+
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+		&tests::sample_review_handoff_marker("main", pr_url, &head_oid),
+	);
+	tests::seed_review_orchestration_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+		&tests::sample_review_orchestration_marker(
+			"main",
+			pr_url,
+			&head_oid,
+			"waiting_for_result",
+			1,
+		),
+	);
+	review_landing_classification_review::record_block_landing_authority_boundary(
+		&state_store,
+		&issue,
+	);
+
+	let mut review_state = tests::sample_pull_request_review_state(
+		pr_url,
+		"main",
+		&head_oid,
+		Some("APPROVED"),
+		"MERGEABLE",
+		"CLEAN",
+		Some("SUCCESS"),
+		0,
+	);
+
+	tests::add_external_review_ack(&mut review_state);
+	tests::add_external_review_pass(&mut review_state);
+	orchestrator::reconcile_post_review_orchestration_with_inspector(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
+	)
+	.expect("post-review orchestration should wait for authority-boundary clearance");
+
+	let marker = tests::persisted_review_orchestration_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+	);
+
+	assert_eq!(marker.phase(), "waiting_for_result");
+	assert!(marker.auto_merge_enabled_at_unix_epoch().is_none());
+	assert!(
+		!invocation_log_path.exists(),
+		"authority-boundary landing requirements must prevent runtime admin merge"
+	);
+	assert!(tracker.comments.borrow().is_empty());
+	assert!(tracker.label_additions.borrow().is_empty());
+}
+
+#[test]
+fn reconcile_post_review_orchestration_blocks_admin_merge_for_human_decision_boundary() {
+	let (temp_dir, config, workflow) = tests::temp_project_layout();
+	let (gh_command_path, invocation_log_path) =
+		tests::install_fake_admin_merge_gh_response(&temp_dir);
+	let config = tests::service_config_with_github_token_env_var_and_command_path(
+		&config,
+		"PATH",
+		&gh_command_path,
+	);
+	let repo_root = config.repo_root().to_path_buf();
+	let issue = review_landing_status_support::post_review_sample_service_owned_issue("In Review");
+	let tracker =
+		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let pr_url = "https://github.com/hack-ink/decodex/pull/173";
+	let merge_subject = r#"{"schema":"decodex/commit/1","summary":"current retained handoff","authority":"PUB-101"}"#;
+	let head_oid =
+		tests::commit_worktree_change(&repo_root, "retained.txt", "ready\n", merge_subject);
+
+	state_store
+		.upsert_worktree("pubfi", &issue.id, "main", &repo_root.display().to_string())
+		.expect("worktree should record");
+
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+		&tests::sample_review_handoff_marker("main", pr_url, &head_oid),
+	);
+	tests::seed_review_orchestration_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+		&tests::sample_review_orchestration_marker(
+			"main",
+			pr_url,
+			&head_oid,
+			"waiting_for_result",
+			1,
+		),
+	);
+	review_landing_classification_review::record_requires_human_authority_boundary(
+		&state_store,
+		&issue,
+	);
+
+	let mut review_state = tests::sample_pull_request_review_state(
+		pr_url,
+		"main",
+		&head_oid,
+		Some("APPROVED"),
+		"MERGEABLE",
+		"CLEAN",
+		Some("SUCCESS"),
+		0,
+	);
+
+	tests::add_external_review_ack(&mut review_state);
+	tests::add_external_review_pass(&mut review_state);
+	orchestrator::reconcile_post_review_orchestration_with_inspector(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
+	)
+	.expect("post-review orchestration should wait for human authority decision");
+
+	let marker = tests::persisted_review_orchestration_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+	);
+
+	assert_eq!(marker.phase(), "waiting_for_result");
+	assert!(marker.auto_merge_enabled_at_unix_epoch().is_none());
+	assert!(
+		!invocation_log_path.exists(),
+		"human authority decisions must prevent runtime admin merge"
+	);
+	assert!(tracker.comments.borrow().is_empty());
+	assert!(tracker.label_additions.borrow().is_empty());
+}
+
+#[test]
+fn reconcile_post_review_orchestration_blocks_admin_merge_for_authority_decision_request() {
+	let (temp_dir, config, workflow) = tests::temp_project_layout();
+	let (gh_command_path, invocation_log_path) =
+		tests::install_fake_admin_merge_gh_response(&temp_dir);
+	let config = tests::service_config_with_github_token_env_var_and_command_path(
+		&config,
+		"PATH",
+		&gh_command_path,
+	);
+	let repo_root = config.repo_root().to_path_buf();
+	let issue = review_landing_status_support::post_review_sample_service_owned_issue("In Review");
+	let tracker =
+		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let pr_url = "https://github.com/hack-ink/decodex/pull/173";
+	let merge_subject = r#"{"schema":"decodex/commit/1","summary":"current retained handoff","authority":"PUB-101"}"#;
+	let head_oid =
+		tests::commit_worktree_change(&repo_root, "retained.txt", "ready\n", merge_subject);
+
+	state_store
+		.upsert_worktree("pubfi", &issue.id, "main", &repo_root.display().to_string())
+		.expect("worktree should record");
+
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+		&tests::sample_review_handoff_marker("main", pr_url, &head_oid),
+	);
+	tests::seed_review_orchestration_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+		&tests::sample_review_orchestration_marker(
+			"main",
+			pr_url,
+			&head_oid,
+			"waiting_for_result",
+			1,
+		),
+	);
+	review_landing_classification_review::record_authority_decision_request(&state_store, &issue);
+
+	let mut review_state = tests::sample_pull_request_review_state(
+		pr_url,
+		"main",
+		&head_oid,
+		Some("APPROVED"),
+		"MERGEABLE",
+		"CLEAN",
+		Some("SUCCESS"),
+		0,
+	);
+
+	tests::add_external_review_ack(&mut review_state);
+	tests::add_external_review_pass(&mut review_state);
+	orchestrator::reconcile_post_review_orchestration_with_inspector(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
+	)
+	.expect("post-review orchestration should wait for authority decision request");
+
+	let marker = tests::persisted_review_orchestration_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&repo_root,
+	);
+
+	assert_eq!(marker.phase(), "waiting_for_result");
+	assert!(marker.auto_merge_enabled_at_unix_epoch().is_none());
+	assert!(
+		!invocation_log_path.exists(),
+		"authority decision requests must prevent runtime admin merge"
+	);
+	assert!(tracker.comments.borrow().is_empty());
+	assert!(tracker.label_additions.borrow().is_empty());
+}
